html/editing/dnd/platform/dragstart-form-associated-custom-element.html

<!doctype html>
<link rel="help" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1935367">
<title>Test dragging works inside form-associated custom elements</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/resources/testdriver-actions.js"></script>

<x-element><span id="contents">Text contents</span></x-element>

<script>
class XElement extends HTMLElement {
  constructor() {
    super();
  }
  static get formAssociated() {
    return true;
  }
}

customElements.define('x-element', XElement);

promise_test(async t => {
  let selection = getSelection();
  t.add_cleanup(() => {
    selection.removeAllRanges();
  });
  let range = new Range();
  range.setStartBefore(contents);
  range.setEndAfter(contents);
  selection.addRange(range);
  let gotDragStart = false;
  contents.addEventListener("dragstart", () => {
    gotDragStart = true;
  });
  await new test_driver.Actions()
    .pointerMove(0, 0, {origin: contents})
    .pointerDown()
    .pointerMove(10, 10, {origin: contents})
    .pointerUp()
    .send();
  assert_true(gotDragStart, 'Should have fired a dragstart event.');
});
</script>
